modificaçoes niveis de permissoes

+++ controllers/product.controller.js
@@ -1,6 +1,7 @@
 const auth = require("../middleware/auth");
 const service = require('../services/role.service');
 const util = require("../util/token");
+const userController = require("../controllers/user.controller");
 const bcrypt = require("bcrypt");
 const { validateUser, validateLogin, validateUserRole } = require("../models/user.model");
 const {Product} = require("../models/product.model")
@@ -8,17 +9,40 @@ const express = require("express");
 const router = express.Router();
 const config = require('config')
 
+function normalizeRBAC(rbac) {
+  let roles = rbac.filter(p => p.product == config.get("product"));
+  if (roles) {
+    roles = roles.map(r => r.roles);
+  }
+  return roles;
+}
+
 router.post("/register", auth, async (req, res) => {
   try {
-    let product = new Product({
-        productname: req.body.productname,
-        amount: req.body.amount,
-        description: req.body.description,
-        price: req.body.price,
-        creationdate: Date(),
-    });
-    let insert = await product.save();
-    res.status(200).send(insert);
+    const _service = new service();
+    await _service.setRole();
+
+    let roles = normalizeRBAC(req.user.rbac); 
+    let resurce = req.path.replace('/', '').split('/').shift(); 
+    let action = "read"; 
+    let context = JSON.parse(req.headers['x-api-context']); 
+    let atrributes = [req.params.atrributes]; 
+
+    let _grant = await _service.grant(roles, action, context, resurce, atrributes);
+    if(_grant){
+      let product = new Product({
+          productname: req.body.productname,
+          amount: req.body.amount,
+          description: req.body.description,
+          price: req.body.price,
+          creationdate: Date(),
+      });
+      let insert = await product.save();
+      res.status(200).send(insert);
+    } else {
+      res.status(400).send({error:'Access denied'});
+    }
+    
   } catch (error) {
     res.status(400).send(error)
   }
@@ -26,15 +50,29 @@ router.post("/register", auth, async (req, res) => {
 
 router.put("/modify/:id", async (req, res) => {
   try {
-    Product.findByIdAndUpdate({_id: req.params.id}, req.body).then(product => {
-      Product.findById({_id: req.params.id}).then(product =>{
-        res.status(200).send(product);
+    const _service = new service();
+    await _service.setRole();
+
+    let roles = normalizeRBAC(req.user.rbac); 
+    let resurce = req.path.replace('/', '').split('/').shift(); 
+    let action = "read"; 
+    let context = JSON.parse(req.headers['x-api-context']); 
+    let atrributes = [req.params.atrributes]; 
+
+    let _grant = await _service.grant(roles, action, context, resurce, atrributes);
+    if(_grant){
+      Product.findByIdAndUpdate({_id: req.params.id}, req.body).then(product => {
+        Product.findById({_id: req.params.id}).then(product =>{
+          res.status(200).send(product);
+        }).catch(err => {
+          res.status(400).send(error);
+        })
       }).catch(err => {
-        res.status(400).send(error);
+        res.send('error: ' + err)
       })
-    }).catch(err => {
-      res.send('error: ' + err)
-    })
+    } else{
+      res.status(400).send({error:'Access denied'});
+    }
   } catch (error) {
     res.status(400).send(error);
   }  
@@ -42,11 +80,26 @@ router.put("/modify/:id", async (req, res) => {
 
 router.patch("/modify/:id", async (req, res) => {
   try {
-    Product.findByIdAndUpdate(req.params.todoId,req.body,{new: true}).then(product => {
-      res.status(200).send(product);
-    }).catch(err => {
-      res.send('error: ' + err)
-    })
+    const _service = new service();
+    await _service.setRole();
+
+    let roles = normalizeRBAC(req.user.rbac); 
+    let resurce = req.path.replace('/', '').split('/').shift(); 
+    let action = "read"; 
+    let context = JSON.parse(req.headers['x-api-context']); 
+    let atrributes = [req.params.atrributes]; 
+
+    let _grant = await _service.grant(roles, action, context, resurce, atrributes);
+    if(_grant){
+      Product.findByIdAndUpdate(req.params.todoId,req.body,{new: true}).then(product => {
+        res.status(200).send(product);
+      }).catch(err => {
+        res.send('error: ' + err)
+      })
+    } else {
+      res.status(400).send({error:'Access denied'});
+    }
+    
   } catch (error) {
     res.status(400).send(error);
   }
@@ -54,11 +107,26 @@ router.patch("/modify/:id", async (req, res) => {
 
 router.delete("/delete/:id", async (req, res) => {
   try {
-    Product.findByIdAndDelete({_id: req.params.id}).then(product => {
-      res.status(200).send('Produto deletado do sistema');
-    }).catch(err => {
-      res.send('error: ' + err)
-    })
+    const _service = new service();
+    await _service.setRole();
+
+    let roles = normalizeRBAC(req.user.rbac); 
+    let resurce = req.path.replace('/', '').split('/').shift(); 
+    let action = "read"; 
+    let context = JSON.parse(req.headers['x-api-context']); 
+    let atrributes = [req.params.atrributes]; 
+
+    let _grant = await _service.grant(roles, action, context, resurce, atrributes);
+    if(_grant){
+      Product.findByIdAndDelete({_id: req.params.id}).then(product => {
+        res.status(200).send('Produto deletado do sistema');
+      }).catch(err => {
+        res.send('error: ' + err)
+      })
+    } else {
+      res.status(400).send({error:'Access denied'});
+    }
+    
   } catch (error) {
     res.status(400).send(error);
   }
@@ -66,15 +134,31 @@ router.delete("/delete/:id", async (req, res) => {
 
 router.get("/list", auth, async (req, res) => {
   try {
-    const page = parseInt(req.query.page) || 1;
-    const limit = parseInt(req.query.limit) || 10;
-    const skip = (page - 1) * limit;
-    let products = await Product.find().skip(skip).limit(limit);
-    if (products == null){
-        return res.status(400).send({error: 'error'});
+    const _service = new service();
+    await _service.setRole();
+
+    let roles = normalizeRBAC(req.user.rbac); 
+    let resurce = req.path.replace('/', '').split('/').shift(); 
+    let action = "read"; 
+    let context = JSON.parse(req.headers['x-api-context']); 
+    let atrributes = [req.params.atrributes]; 
+
+    let _grant = await _service.grant(roles, action, context, resurce, atrributes);
+
+    if(_grant){
+      const page = parseInt(req.query.page) || 1;
+      const limit = parseInt(req.query.limit) || 10;
+      const skip = (page - 1) * limit;
+      let products = await Product.find().skip(skip).limit(limit);
+      if (products == null){
+          return res.status(400).send({error: 'error'});
+      } else {
+          res.status(200).send({status: 'success', products});
+      }
     } else {
-        res.status(200).send({status: 'success', products});
-    } 
+      res.status(400).send({error:'Access denied'});
+    }
+     
   } catch (error) {
     res.status(400).send(error);
   }
